+++ config/push.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    /*
+     * The models for rating tables.
+     */
+    'push_server' => env('PUSH_SERVER', ''),
+    'push_domain_header' => env('PUSH_DOMAIN_HEADER', ''),
+    'push_token_header' => env('PUSH_TOKEN_HEADER', ''),
+    'push_domain' => env('PUSH_DOMAIN', ''),
+    'push_token' => env('PUSH_TOKEN', ''),
+];